convert 64 bit height usages in rpcs tp 32 bit. add univalue contructor that take uint32_t

+++ src/rpc/mining.cpp
@@ -587,7 +587,7 @@ static UniValue getblocktemplate(const JSONRPCRequest& request)
     result.pushKV("sizelimit", (int64_t)GetCurrentMaxBlockSize());
     result.pushKV("curtime", pblock->GetBlockTime());
     result.pushKV("proof", HexStr(pblock->proof));
-    result.pushKV("height", (int64_t)(pindexPrev->nHeight+1));
+    result.pushKV("height", (pindexPrev->nHeight+1));
 
     if (!pblocktemplate->vchCoinbaseCommitment.empty() && fSupportsSegwit) {
         result.pushKV("default_witness_commitment", HexStr(pblocktemplate->vchCoinbaseCommitment.begin(), pblocktemplate->vchCoinbaseCommitment.end()));
